Cache panels on init and layout change
